master <> develop (#1177)
* fix: add sql endpoint auth + make `run_query` fail gracefully (#1136)

* fix: add sql endpoint auth

* retrigger ci?

---------


* chore: switch to manual sqlx query logging (#1138)

* chore: switch to manual sqlx query logging

* fix test

* uat

* add rustdoc

---------


* fix: update readme (#1141)

* enhancement: make `NODE_BLOCK_PAGE_SIZE` configurable (#1147)

* Replace blocking `reqwest` usage with async (#1148)

* Revert "Replace blocking `reqwest` usage with async" (#1151)

Revert "Replace blocking `reqwest` usage with async (#1148)"

This reverts commit 19dbf2acfb5f8c3c8509ec48617d9412868769d0.

* docs: adjust docs to include `--block-page-size` on start commands (#1153)

Adjust docs for start commands

* Replace blocking `reqwest` usage with async (#1152)

* fix: don't skip blocks on constraint error (#1154)

* fix: don't skip blocks on constraint error

* fix test

---------


* Adding resources to helm chart (#1160)

* Update --replace-indexer and add --remove-data (#1146)

* remove debug info message (#1163)

* docs: update non-book docs (#1161)

* docs: update non-book docs

* updates

* markdownrc

* update contributing.md

* update release schedule

* update example readmes

---------


* Fix incorrect schema digest (#1165)

* forc index deploy: register correct schema

* bugfix: we are removing the schema and registering it again

* log registered asset version

* add docs CI (#1158)

* enhancement: implement many-to-many relationships (#1140)

* enhancement: impl m2m inserts

* update implementation

* get the explorer working

* update native exec

* cleanup + updates

* get union types working

* use multiple queries

* rebase origin/develop

* fix test

---------


* chore: improve UX of `forc index` plugin commands (#1142)

* Remove forc index new

* Allow for optional namespace in forc index init

* Remove profile and target flags from forc index build; change release bool to debug

* Remove service check from forc index check

* Move service check into forc index status

* Cleanup

* Remove short arg collision from forc index postgres

* Fix CI

* Replace forc_index_init with forc_index_new

* Adjust tests and documentation

* Address feedback; clean up some other things

* Address everything for real this time

* ci: streamline develop <> master merges (#1166)

* ci: use new formatter + fix master<>develop merging  (#1169)

fix: new ci formatting + fix master<>develop merge


* v0.19.0 (#1172)


---------

+++ packages/fuel-indexer-database/postgres/src/lib.rs
@@ -24,7 +24,8 @@ use fuel_indexer_macro_utils::metrics;
 
 use chrono::{DateTime, NaiveDateTime, Utc};
 
-const NONCE_EXPIRY: u64 = 3600; // 1 hour
+/// The nonce expiry time in seconds.
+const NONCE_EXPIRY_SECS: u64 = 3600;
 
 #[cfg_attr(feature = "metrics", metrics)]
 pub async fn put_object(
@@ -67,7 +68,10 @@ pub async fn run_query(
         .fetch_all(conn)
         .await?
         .iter()
-        .map(|r| r.get::<JsonValue, usize>(0))
+        .filter_map(|r| match r.try_get::<JsonValue, usize>(0) {
+            Ok(v) => Some(v),
+            Err(_e) => None,
+        })
         .collect())
 }
 
@@ -571,8 +575,9 @@ pub async fn register_indexer_asset(
         .await?;
 
     info!(
-        "Registered Asset({:?}) to Indexer({}).",
+        "Registered Asset({:?}) with Version({}) to Indexer({}).",
         asset_type,
+        digest,
         index.uid()
     );
 
@@ -771,6 +776,7 @@ pub async fn remove_indexer(
     conn: &mut PoolConnection<Postgres>,
     namespace: &str,
     identifier: &str,
+    remove_data: bool,
 ) -> sqlx::Result<()> {
     execute_query(
         conn,
@@ -795,16 +801,6 @@ pub async fn remove_indexer(
     execute_query(
         conn,
         format!(
-            "DELETE FROM index_asset_registry_schema WHERE index_id IN
-            (SELECT id FROM index_registry
-                WHERE namespace = '{namespace}' AND identifier = '{identifier}')"
-        ),
-    )
-    .await?;
-
-    execute_query(
-        conn,
-        format!(
             "DELETE FROM index_registry WHERE id IN
             (SELECT id FROM index_registry
                 WHERE namespace = '{namespace}' AND identifier = '{identifier}')"
@@ -815,40 +811,52 @@ pub async fn remove_indexer(
     execute_query(
         conn,
         format!(
-            "DELETE FROM graph_registry_columns WHERE type_id IN (SELECT id FROM graph_registry_type_ids WHERE schema_name = '{namespace}' AND schema_identifier = '{identifier}');"
+            "DELETE FROM index_asset_registry_schema WHERE index_id IN
+            (SELECT id FROM index_registry
+                WHERE namespace = '{namespace}' AND identifier = '{identifier}')"
         ),
     )
     .await?;
 
-    execute_query(
-        conn,
-        format!(
-            "DELETE FROM graph_registry_type_ids WHERE schema_name = '{namespace}' AND schema_identifier = '{identifier}';"
-        ),
-    )
-    .await?;
+    if remove_data {
+        execute_query(
+            conn,
+            format!(
+                "DELETE FROM graph_registry_columns WHERE type_id IN (SELECT id FROM graph_registry_type_ids WHERE schema_name = '{namespace}' AND schema_identifier = '{identifier}');"
+            ),
+        )
+        .await?;
 
-    execute_query(
-        conn,
-        format!(
-            "DELETE FROM graph_registry_root_columns WHERE root_id = (SELECT id FROM graph_registry_graph_root WHERE schema_name = '{namespace}' AND schema_identifier = '{identifier}');"
-        ),
-    )
-    .await?;
+        execute_query(
+            conn,
+            format!(
+                "DELETE FROM graph_registry_type_ids WHERE schema_name = '{namespace}' AND schema_identifier = '{identifier}';"
+            ),
+        )
+        .await?;
 
-    execute_query(
-        conn,
-        format!(
-            "DELETE FROM graph_registry_graph_root WHERE schema_name = '{namespace}' AND schema_identifier = '{identifier}';"
-        ),
-    )
-    .await?;
+        execute_query(
+            conn,
+            format!(
+                "DELETE FROM graph_registry_root_columns WHERE root_id = (SELECT id FROM graph_registry_graph_root WHERE schema_name = '{namespace}' AND schema_identifier = '{identifier}');"
+            ),
+        )
+        .await?;
 
-    execute_query(
-        conn,
-        format!("DROP SCHEMA {namespace}_{identifier} CASCADE"),
-    )
-    .await?;
+        execute_query(
+            conn,
+            format!(
+                "DELETE FROM graph_registry_graph_root WHERE schema_name = '{namespace}' AND schema_identifier = '{identifier}';"
+            ),
+        )
+        .await?;
+
+        execute_query(
+            conn,
+            format!("DROP SCHEMA {namespace}_{identifier} CASCADE"),
+        )
+        .await?;
+    }
 
     Ok(())
 }
@@ -882,7 +890,7 @@ pub async fn create_nonce(conn: &mut PoolConnection<Postgres>) -> sqlx::Result<N
         .unwrap()
         .as_secs();
 
-    let expiry = now + NONCE_EXPIRY;
+    let expiry = now + NONCE_EXPIRY_SECS;
 
     let row = sqlx::QueryBuilder::new(&format!(
         "INSERT INTO nonce (uid, expiry) VALUES ('{uid}', {expiry}) RETURNING *"
@@ -971,4 +979,13 @@ pub async fn indexer_owned_by(
     }
 
     Err(sqlx::Error::RowNotFound)
+}
+
+#[cfg_attr(feature = "metrics", metrics)]
+pub async fn put_many_to_many_record(
+    conn: &mut PoolConnection<Postgres>,
+    query: String,
+) -> sqlx::Result<()> {
+    execute_query(conn, query).await?;
+    Ok(())
 }
